Write City tests
